backend/segmentation/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Deep learning frameworks
torch==2.1.1
torchvision==0.16.1

# Image processing
opencv-python-headless==4.8.1.78
pillow==10.1.0
scikit-image==0.22.0

# Scientific computing
numpy==1.24.3
scipy==1.11.4

# Data handling
pydantic==2.5.0

# Logging and monitoring
python-json-logger==2.0.7

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0

# Production deployment
gunicorn==21.2.0